+++ src/EventSystem/NormalizedValueToRawValue/index.ts
@@ -1,0 +1,3 @@
+export * as Event from './Event';
+export * as EventListener from './EventListener';
+export * from './NormalizedValueToRawValueEventManager';