+++ src/index.ts
@@ -1,0 +1,11 @@
+import { Container } from 'typedi';
+
+export * as BrowserEvent from './BrowserEvent';
+export * as Endpoint from './Endpoint';
+export * as Error from './Error';
+export * as EventSystem from './EventSystem';
+export * as Factory from './Factory';
+export * as Service from './Service';
+export * as Type from './Type';
+export * from './EmberNexus';
+export { Container };